requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alphashape==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
chardet==6.0.0.post1
charset-normalizer==3.4.4
click==8.3.1
click-log==0.4.0
Crawl4AI==0.8.0
cryptography==46.0.5
cssselect==1.4.0
distro==1.9.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastuuid==0.14.0
filelock==3.24.3
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.3.1
litellm==1.81.14
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
networkx==3.6.1
nltk==3.9.2
numpy==2.4.2
openai==2.21.0
packaging==26.0
patchright==1.58.0
pillow==12.1.1
playwright==1.58.0
propcache==0.4.1
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
pyOpenSSL==25.3.0
python-dotenv==1.2.1
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rtree==1.4.1
scipy==1.17.0
shapely==2.1.2
shellingham==1.5.4
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.8.3
tf-playwright-stealth==1.2.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
trimesh==4.11.2
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
